modal_analysis/requirements.txt
ansys-dpf-core
PyQt5
numpy
psutil
pytest
